Refactor handling of Trials tasks
Ensure trials are called "Interactor Trials" and "Observer Trials". Remove duplicate code by creating base classes for the training & main Trials tasks, and moving all Trials classes to a single file ("task_trials.py").
