+++ Client/standardised_values.py
@@ -7,8 +7,8 @@ SERVER_NAME = 'SubtleGame'
 TASK_SANDBOX = 'sandbox'
 TASK_NANOTUBE = 'nanotube'
 TASK_KNOT_TYING = 'knot-tying'
-TASK_TRIALS = 'trials'
-TASK_TRIALS_TRAINING = 'trials-training'
+TASK_TRIALS_INTERACTOR = 'trials'
+TASK_TRIALS_INTERACTOR_TRAINING = 'trials-training'
 TASK_TRIALS_OBSERVER = 'trials-observer'
 TASK_TRIALS_OBSERVER_TRAINING = 'trials-observer-training'
 # Simulation
@@ -96,9 +96,9 @@ PLAYER_FINISHED = 'Finished'
 SHARED_STATE_KEYS_AND_VALS = {
     KEY_MODALITY: [MODALITY_HANDS, MODALITY_CONTROLLERS],
     KEY_GAME_STATUS: [WAITING, IN_PROGRESS, FINISHED],
-    KEY_ORDER_OF_TASKS: [TASK_NANOTUBE, TASK_KNOT_TYING, TASK_TRIALS, TASK_TRIALS_TRAINING, TASK_TRIALS_OBSERVER,
+    KEY_ORDER_OF_TASKS: [TASK_NANOTUBE, TASK_KNOT_TYING, TASK_TRIALS_INTERACTOR, TASK_TRIALS_INTERACTOR_TRAINING, TASK_TRIALS_OBSERVER,
                          TASK_TRIALS_OBSERVER_TRAINING],
-    KEY_CURRENT_TASK: [TASK_SANDBOX, TASK_NANOTUBE, TASK_KNOT_TYING, TASK_TRIALS, TASK_TRIALS_TRAINING,
+    KEY_CURRENT_TASK: [TASK_SANDBOX, TASK_NANOTUBE, TASK_KNOT_TYING, TASK_TRIALS_INTERACTOR, TASK_TRIALS_INTERACTOR_TRAINING,
                        TASK_TRIALS_OBSERVER, TASK_TRIALS_OBSERVER_TRAINING],
     KEY_TASK_STATUS: [READY, IN_PROGRESS, FINISHED],
     KEY_TRIALS_TIMER: [STARTED, FINISHED],
